soc: arm64: qemu_cortex_a53: fix soc.h usage
<soc.h> has been traditionally been used as a proxy to HAL headers,
register definitions, etc. Nowadays, <soc.h> is anarchy. It serves a
different purpose depending on the SoC. In some cases it includes HALs,
in some others it works as a header sink/proxy (for no good reason), as
a register definition when there's no HAL... To make things worse, it is
being included in code that is, in theory, non-SoC specific.

This patch is part of a series intended to improve the situation by
removing <soc.h> usage when not needed, and by eventually removing it.

(cherry picked from commit 22dfbca6f0d646ecb26a5739dbf60368eee625bd)

Original-Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: 22dfbca6f0d646ecb26a5739dbf60368eee625bd
Change-Id: Ia8222d3535ab67a58053b345a9694f6d25a1447d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3691584
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>