soc: arm64: intel_socfpga: fix soc.h usage
<soc.h> has been traditionally been used as a proxy to HAL headers,
register definitions, etc. Nowadays, <soc.h> is anarchy. It serves a
different purpose depending on the SoC. In some cases it includes HALs,
in some others it works as a header sink/proxy (for no good reason), as
a register definition when there's no HAL... To make things worse, it is
being included in code that is, in theory, non-SoC specific.

This patch is part of a series intended to improve the situation by
removing <soc.h> usage when not needed, and by eventually removing it.

In this case, the file did not use anything defined in <soc.h>, it only
required access to integer types, e.g. uint32_t.

(cherry picked from commit e74c0838144338da36f1048aa0b2254f1ed08fa7)

Original-Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: e74c0838144338da36f1048aa0b2254f1ed08fa7
Change-Id: I2ff9d3607813af1218c49d5a77d768b0fdf67fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3691578
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>