drivers: serial: ns16550: use MMIO device depending on Kconfig option
Add a new selectable Kconfig option to decide wether the device driver
is a MMIO device or not. Previous to this patch, the decision was maded
based on the existence of a definition in <soc.h>. The design was
fragile, as code compiled anyway if the definition was not present.

All platforms/boards that had the definition in <soc.h> select the
Kconfig option in their respective defconfig files.

(cherry picked from commit 92f488497fb5bb7cf40fd027cae2a97d426b16a1)

Original-Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: 92f488497fb5bb7cf40fd027cae2a97d426b16a1
Change-Id: Ieabf78df3398cda75a4f5fb822df233d6b32fe63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3691560
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>