cmake: use all compile options arguments for uniqueness
zephyr_library_compile_options() was attempting to create
unique hashes for compile options in order to prevent creating
them multiple times.  However, it was only using the first
argument to create the hash, so if multiple libraries had
different compile options but the first line was the same,
the second set would be mistaken for the first set and would
actually be passed the first set during compilation instead
of its own set.

The fix should be to use the entire compile options argument
list to create the hash so they should only match if the entire
options list is exactly the same.

This is a continued fix for #43835

(cherry picked from commit 3da90ad21ffc2fa102d936cd61809af7bb6cd2ab)

Original-Signed-off-by: Nicholas Lowell <nlowell@lexmark.com>
GitOrigin-RevId: 3da90ad21ffc2fa102d936cd61809af7bb6cd2ab
Change-Id: I38f738c56f1ddd9d41a6e5155d66b936694ae010
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3698022
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>