Bluetooth: ISO: Modify cc in bt_iso_chan_path
The CC/Codec Configuration in bt_iso_chan_path was
defined as an array of size 0. This meant that the
CC always had to be allocated right after the
bt_iso_chan_path struct.

This does not give a very flexible API, and also makes
it impossible for two bt_iso_chan_path to share the same
CC.

The API is modified so that the CC is simply a pointer
to a an array now.

(cherry picked from commit 10990302101bd9b45024b91420ddbed7c58201ea)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 10990302101bd9b45024b91420ddbed7c58201ea
Change-Id: I7443141ea1f32fc5115bc24ff79dcbaa41b7f178
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3873353
Tested-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>