tests/intel_adsp/smoke: Revert WAIT_FOR breakage
The evolution of the original WAIT_FOR trick in this test, first to
move it to a generic spot and then to make it take a delay
paramatrization, have had the unfortunate side effect of breaking the
original usage.

And... it's really not following the design intent anyway.  The idea
here was to spin hard waiting on external hardware (a python script
running on the host) to deliver an interrupt.  Both of the new tricks
(suspending the thread with a sleep and hammering the HDA wall clock
for time) can affect interrupt delivery.

Just restore the original implementation (albeit with a new name,
"AWAIT", since the original was stolen) until someone has a chance to
make this work properly with the new API.  There's no reason it can't,
it just doesn't yet.

Fixes #43828

(cherry picked from commit 5139d052d87bc8571c20aedda61796c4735f756c)

Original-Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 5139d052d87bc8571c20aedda61796c4735f756c
Change-Id: I547299a2848a11ad2540c3c0996306613f671935
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3695502
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>