samples/net/virtual: Increase sprintk buffers to hold parameter types
This sample app uses a stack buffer to hold the interface name which is
generated at runtime using sprintk. The buffer is only sized for the
expected range of values, not the full possible range of the datatype
(int).

To mitigate this, increase the size of the buffer to hold the full range of
an int value.

(cherry picked from commit be606f09326ae793363c04fe8ae2c2f829d20d6b)

Original-Signed-off-by: Keith Packard <keithp@keithp.com>
GitOrigin-RevId: be606f09326ae793363c04fe8ae2c2f829d20d6b
Change-Id: Id3a53d06ba201174c2685db236231f3c47078ec6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3698294
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>