mgmt/osdp: discard secure channel if KEYSET is ACKed in plaintext
HID readers are responding to a KEYSET command with an ACK in plaintext
instead of using the current session keys to acknowledge this command
(which is the reasonable thing to do as the command itself was received
encrypted with the old key). Since the spec doesn't say anything about
this, both methods are technically correct.

Make changes to CP so it allows ACKs in plaintext for KEYSET command in
particular and make the PD implementation do this too as it makes the code
look cleaner there (perhaps HID did it for the same reasons :D).

(cherry picked from commit 0effd82b7ad546605e10f0d516f5ff4cc9ffaa88)

Original-Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>
GitOrigin-RevId: 0effd82b7ad546605e10f0d516f5ff4cc9ffaa88
Change-Id: Ia4654b1b541f825a9c532d8dcbb056dd9b96a603
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4410679
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>