dts: riscv: add #address-cells to all interrupt controllers
This mirrors #36499 and other PRs that added them for other
architectures.

This silences a large number of dtc warnings due to the missing
property. It seems reasonable to require an address-cells property since
any interrupt controller could be the parent of an interrupt-map.

The only device actually using interrupt-maps is neorv32, and it needs
an address-cells of 2 (since this is the default if none is specified it
worked like that before this change).

While I touched this, I reordered the properties for consistency across
boards, but there's a lot of variance here already.

(cherry picked from commit 07ac630281c353c03b2b147f69dc8ff30a0fc357)

Original-Signed-off-by: Olof Johansson <olof@lixom.net>
GitOrigin-RevId: 07ac630281c353c03b2b147f69dc8ff30a0fc357
Change-Id: I12022e9131cad2d8985446d68d78921cc2d6e6aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3745820
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>