Bluetooth: Host: Refactor `bt_setup_public_id_addr`
Move the flagging of `BT_DEV_STORE_ID` to a separate if-statement.

This clarifies the meaning of the code, and paves way for wrapping the
if-statement with the call to `bt_smp_irk_get` in another conditional.

(cherry picked from commit 058dc7c0bcf3ab45c758e11d4f87d12186625790)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 058dc7c0bcf3ab45c758e11d4f87d12186625790
Change-Id: I991c19493b181701e724af94142baef2ca42e528
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3841966
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>