drivers: usb: device: stm32: only disable the ULPI clock for USB FS
On SoCs with two USB controller, Zephyr selects the FS one, and in that
case the ULPI clock needs to be disabled in sleep mode for the
controller to work.

On SoCs with a single USB HS controller, this operation is not needed.

(cherry picked from commit b2f969e5627fad419feb7e1895178a4508ddf186)

Original-Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
GitOrigin-RevId: b2f969e5627fad419feb7e1895178a4508ddf186
Change-Id: If3dcaa61fee18557265f48fa813f5997bb1a3f97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3724726
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>