tests: bluetooth: bsim: Pretend that the controller supports CIS
The test relies on enabling connection iso features in the host. As the
controller does not yet support them, this configuration does not really
make sense. But because the test is not using this functionality, we
work around it by pretending the controller supports it.

(cherry picked from commit bdc19bf2747070fcf930b2815e8a5279d30f8605)

Original-Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>
GitOrigin-RevId: bdc19bf2747070fcf930b2815e8a5279d30f8605
Change-Id: I9779b5ba57d5f5e9bd04adb0430ee0430dd6bdf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3712960
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>