dts: arm: st: h7: add max-erase-time property to the flash node
The erase time for a sector differs by a few order of magnitude for the
various SoCs of the STM32H7 family. Declare it in the device tree using
the existing max-erase-time property. For that it is necessary to add
"st,stm32-nv-flash" to the compatible node.

(cherry picked from commit 98a02fefc740f4f7e63787deb7fdfb906bdb9c63)

Original-Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
GitOrigin-RevId: 98a02fefc740f4f7e63787deb7fdfb906bdb9c63
Change-Id: I616d98091b8df8e0e3a0788c82f9762840e7eddf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3716491
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>