mgmt/osdp: Add support for event delivery and notifications
The CP app sends PD a "command" and the PD responds to it. Some times,
the PD has something that it wants to tell the PD which it does so in
response to POLL command. Both CP and PD apps need a way to exchange
these info over the OSDP bus. To archive this we will introduce what are
called "events" that allow the PD app to enqueue and CP app to get
notified.

This is analogous to the incumbent "commands" abstraction where, the CP
app enqueues a command and the PD app gets notified of it.

(cherry picked from commit 73809472f8866acdbc036b6979b61ae7f10f02e9)

Original-Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>
GitOrigin-RevId: 73809472f8866acdbc036b6979b61ae7f10f02e9
Change-Id: I183393f3fe7b7023d3b37b1f06af637f22072579
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4163132
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>