net: lwm2m: Try to reuse registration in case of a network error
For most of the times it is not necessary to do a full registration
once a connection is established after a network error. This is in
particular not needed if lifetime is not yet expired and the server
does not refuse a registration update.

(cherry picked from commit 7532a23dc9e50aadb3c74bae408abb1bc8c4dbc2)

Original-Signed-off-by: Andreas Chmielewski <andreas.chmielewski@grandcentrix.net>
GitOrigin-RevId: 7532a23dc9e50aadb3c74bae408abb1bc8c4dbc2
Change-Id: Iaa132a024dc56e8c3e5f0e3f9a1d4fb9bee29f58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3815047
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>