cmake: initial sysbuild / multi image support
This is the initial commit with system build, sysbuild.

Using CMake as infrastructure together with the Zephyr sysbuild allows
us to support a convenient way of building a sample and allow for extra
images to be built as part of a larger system.

It uses Kconfig for configuration of image builds.
This allows for future extension with additional build images.

(cherry picked from commit 9d6cc39d6fcdaa79fbb19ef6dc16f40422bdbebd)

Original-Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
GitOrigin-RevId: 9d6cc39d6fcdaa79fbb19ef6dc16f40422bdbebd
Change-Id: I06a723473cad3dbf3503126419a84790c8337895
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3810898
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>