lib: os/cbprintf: make char[] work in C++ and tagged arguments
Strings in C++ are usually of type char[] instead of char*.
Therefore, this needs to be taken into account when doing
tagged arguments or else char[] would be tagged as simple
pointer. So fix this by adding a way to get the type of
elements in an array to determine if it is a char array.

Fixes #46761

(cherry picked from commit 73a940e3b7c60af0e963107b459a62a9a1651d72)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 73a940e3b7c60af0e963107b459a62a9a1651d72
Change-Id: I66d22719108a12afc8738357c81fc3ac852dc278
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3737845
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>