drivers: timer: stm32 lptim fix long time interrupt locking
The waiting of the flag ARROK could be quite long (100µs-200µs in my
test). During this time, the interrupts are locked that is not
recommended. To avoid this, we manage the update of the autoreload value
in interruption

(cherry picked from commit a703cbe9ce1315f2460d86de7ee8a55190cf7995)

Original-Signed-off-by: Julien D'Ascenzio <julien.dascenzio@paratronic.fr>
GitOrigin-RevId: a703cbe9ce1315f2460d86de7ee8a55190cf7995
Change-Id: I144e8720ab1ad582b09b88c1eb8e44d418956e57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3822989
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>