tests/kernel/context: Fixup slop in cpu_idle test case
This test was written to idle for exactly 1ms and wake up with zero
error, which is just too tight for some platforms (and worked on
emulators where the tick rate is 10x coarser only because 0 == 0!).

And it's not clear that it's testing anything we promise in
documentation, regardless.  Early wakeups are not an error and
absolutely not disallowed, yet the test is treating the wakeup like a
sleep.

Clean it up a bit and relax the tolerance to what we can compute
host quantum for emulators), and allow 1 tick of slop on either side to
permit slightly early wakeups while still verifying that "yes, the idle
did idle".

Fixes #46641

(cherry picked from commit b65277990a29a3e3a5f9852ddb4be8a60837990a)

Original-reliably: do all the math in ticks, idle for 10ms (i.e. longer than a
Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: b65277990a29a3e3a5f9852ddb4be8a60837990a
Change-Id: I0b7c6cd63b322bd50bc7f5ea28e09f2fe1c7d027
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3740822
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>