drivers: spi: spi-bitbang: Add support for half duplex
Implement support for half duplex communication in the bit bang SPI
driver. The SPI driver will use the MOSI pin is for both TX and RX
operations when using half duplex mode.

In half-duplex mode, the driver configures the MOSI pin as an input
pin for input only transactions. Transactions that are bidirectional
are forbidden. After an SPI transaction, the MOSI pin is left as an
input if it was an RX transaction or an output after a TX
transaction. Like before, the MOSI pin is initialized as an
(inactive) output pin when the bus is initialized.

(cherry picked from commit d33b89befe89b50e3e7cafc81028da6cb6d35075)

Original-Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>
GitOrigin-RevId: d33b89befe89b50e3e7cafc81028da6cb6d35075
Change-Id: I47c5050bc94fd30f18f279e24fe234a3b836f5cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3760516
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>