linker: Fix _vector_end symbol location
The 219d5b5adb22 commit is breaking the STM32F0 platforms.

The problem is that the PR is changing the location the _vector_end
symbol is pointing to. These platforms require the symbol to be pointing
to a location in memory after the IRQ vector table.

Fix the issue by introducing a new _vector_end symbol when the IRQ
vector table is present that points to the correct location.

(cherry picked from commit 8c4f98d263865d00199fc9b5f7870a0f134c634b)

Original-Signed-off-by: Carlo Caione <ccaione@baylibre.com>
GitOrigin-RevId: 8c4f98d263865d00199fc9b5f7870a0f134c634b
Change-Id: Iaed7c26161a3282a27a179c8a023c25350b4592a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3748255
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>