drivers: adc: stm32f1 adc driver enable peripheral
This commit avoid re-enabling the ADC,
because this starts conversion.
According to the RefMan (RM0008) of the stm32F10x mcu,
enabling the ADC will start the conversion
if the ADC is already enabled. "Conversion starts when
this bit holds a value of 1 and a 1 is written to it."
That's not what we want.

(cherry picked from commit 085cdd927733d6856573b1c2432b9e0fb9009ea6)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: 085cdd927733d6856573b1c2432b9e0fb9009ea6
Change-Id: Ia18f5cc7791974921841bcd4644c944c2867a47c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3755390
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>