boards: nrf5340dk_nrf5340: Clean up common cpuapp dts node definitions
In commit 619a498845fa33fb830215b4caca21d8438a5d90, the initialization
of the spi2 node has been moved from the nrf5340_cpuapp_common.dts file
to separate dts files for secure/non-secure targets. That was done
to make the node unavailable in the nRF5340 PDK, that was also using
the common file, because it was non-functional on that board.
For the same reason, the initialization of the usbd node could not
be done in the common file and instead it had to be added separately
for the secure and non-secure targets.
Now that the nRF5340 PDK is no longer supported, there is no point in
keeping such complication. The usbd node initialization is moved to the
common file but the spi2 node initialization is removed and instead the
spi4 node is enabled, as it is labeled as arduino_spi and is assigned
with the same pins.

(cherry picked from commit 915419ca2b3eab82e76ce8c24f8aff43ffbd971d)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 915419ca2b3eab82e76ce8c24f8aff43ffbd971d
Change-Id: I36afe24057b72168af90b6559226f8d54e6af982
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3755396
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>