scripts: gen_defines: fix tokenization of array elements
This fixes commit
3d5cc38cf6f58b363828059fa11e151112e9cce0.

That commit only works if removing the quotes from e.g. a string in an
array actually results in a token. If the string's value is something
like

  "foo,bar"

though, it will result in the "token"

  foo,bar

in the generated output.

This is wrong; fix it using the new edtlib.str_as_token() API
introduced to allow callers to perform the same procedure as that
library for fixing up strings.

(cherry picked from commit d6f68f0bbb5187371c50205cad8afe8318366cd9)

Original-Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: d6f68f0bbb5187371c50205cad8afe8318366cd9
Change-Id: Ia61ab9e24efa1710482eeadf3e1de1296838542b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3772311
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>