scripts: gen_handles: output dependency graph
Output the final dependency graph as a `.dot` file, which when rendered
by graphviz can be easier to comprehend than the text descriptions.

This output is optional in that it will not be generated if `graphviz`
is not installed.

(cherry picked from commit 29942475c5ce4c9f1d7a76273912777445f7d429)

Original-Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: 29942475c5ce4c9f1d7a76273912777445f7d429
Change-Id: I72f746c94c49622d91b9cef1760ee30f38202727
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3797361
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>