fcb_settings: Use weak function to get storage flash area
On MCU with bank swap capabilities, the offset of the storage area is
not the same before and after a bank swap. This commit introduce a weak
function which returns the default flash area of the storage partition.
On MCU with the bank swap capability the user can define its own
function to get the proper flash area depending on which bank the fw is
run from.
This commit is a workaround for #47732

(cherry picked from commit 80c97c37de4e0db35da6c3c66fc23a40646f2165)

Original-Signed-off-by: Nicolas VINCENT <nicolas.vincent@vossloh.com>
GitOrigin-RevId: 80c97c37de4e0db35da6c3c66fc23a40646f2165
Change-Id: I13e8e39e148ad669566aecd4deda0bf86b164a2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3841946
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>