ipc_service: Add ipc_rpmsg_deinit function
This function does the opposite of the ipc_rpmsg_init function and
allows the rpmsg instance to be torndown in the case that the application
wishes to do so. It is intended to be used by the ipc_service subsystem.

(cherry picked from commit f16a19c33f9f94075db75788483cbe8232bcc5e9)

Original-Signed-off-by: Jackson Cooper-Driver <jackson.cooper---driver@amd.com>
GitOrigin-RevId: f16a19c33f9f94075db75788483cbe8232bcc5e9
Change-Id: I809ad623b50df5883d01caf141e0842555dce746
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3932761
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>