drivers: dma stm32 keep the channel busy if overriden
When the channel is overriden by the HAL,
the dma irq must always be handled (even if not busy).
Only when the dma channel is not overriden by the HAL
the irq is exiting when the channel is no more busy.

(cherry picked from commit 5412c2126da54b3e11c32311cd263bf1eaa7396a)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: 5412c2126da54b3e11c32311cd263bf1eaa7396a
Change-Id: I5702deb78d794bf631237ca564ccd836a4d85a8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3788355
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>