drivers: ssd16xx: Add an API to read raw RAM contents
Add the ssd16xx_read_ram() function that can be used to read the raw
contents of the two display RAM in the controller. This function is
similar to display_read(), but lets the caller specify the RAM to
operate on. The intention is that this can be used to read out the
contents of the old frame buffer after a partial refresh that
automatically swaps the black/red buffers.

(cherry picked from commit 61306aedb405ff715be3f7c3f7803dd5e2c0f28b)

Original-Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>
GitOrigin-RevId: 61306aedb405ff715be3f7c3f7803dd5e2c0f28b
Change-Id: Iabccd5ecf02021e671e626c0e6e6147fe3576fdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3791136
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>