drivers: gd7965: Remove the need to use heap memory
The GD7965 driver uses malloc to allocate a single line buffer when
clearing the display. This buffer is used to clear the display line by
line. This as two problems. First, it means that the driver introduces
an unnecessary requirement to support heap allocations. Second, it
causes a lot of weird and unnecessary SPI transactions that look like
partial updates without the actual refresh. This is very inefficient
since the same action can be performed in a single SPI transaction.

Add a gd7965_write_cmd_pattern() helper function that writes a pattern
of a specified length to a register in the device.

(cherry picked from commit 376973423f619f6081afdb9a0a215c32ded1e7c9)

Original-Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>
GitOrigin-RevId: 376973423f619f6081afdb9a0a215c32ded1e7c9
Change-Id: I2a35a5da91bb4e8803886b0a0843abfd1319d703
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3776867
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>