dts: bindings: Remove compatible from nxp,imx-flexspi-device.yaml
The 'nxp,imx-flexspi-device' compatible that was specified in
the YAML file is not an actual compatible that is intented to be
used.  Instead the YAML file is included in other YAML files
that set specific compatible.

If the 'nxp,imx-flexspi-device' was used we'd actually get an error
from edtlib because of having the same compatible specified by
two different YAML files.

(cherry picked from commit 91feb4ce17b697fbb312aae2065d0909ff03f84d)

Original-Signed-off-by: Kumar Gala <galak@kernel.org>
GitOrigin-RevId: 91feb4ce17b697fbb312aae2065d0909ff03f84d
Change-Id: I34b1686e374150e66e83dd337194ee1b3163aa35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3768204
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>