Bluetooth: doc: Add warning for bt_data_parse
The fact that the function consumed its argument was not clear previously.

Fixes #47676.

The alternative is to change the API to not consume the data, but as it's
an API change, I suggest we only add the warning.

(cherry picked from commit f8531fd3790c9774cd00865e4ab8c465fd408382)

Original-Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>
GitOrigin-RevId: f8531fd3790c9774cd00865e4ab8c465fd408382
Change-Id: I5524882b35c56feced8dedf4f5659a694035e37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3768212
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>