drivers: can: mcan: acknowledge all received frames
The Bosch M_CAN IP does not support RX filtering of the RTR bit, so the
driver handles this bit in software.

If a recevied frame matches a filter with RTR enabled, the RTR bit of the
frame must match that of the filter in order to be passed to the RX
callback function. If the RTR bits do not match the frame must be dropped.

Improve the readability of the the logic for determining if a frame should
be dropped and add a missing FIFO acknowledge write for dropped frames.

(cherry picked from commit 5e74f722206ad5873d4024d90b59905771e80322)

Original-Fixes: #47204
Original-Signed-off-by: Henrik Brix Andersen <hebad@vestas.com>
GitOrigin-RevId: 5e74f722206ad5873d4024d90b59905771e80322
Change-Id: Ic7abebc811ca9f3dedac6ab2e8418ee26a2139e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3772265
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>