drivers: can: mcux: flexcan: fix handling of RTR frames
When installing a RX filter, the driver uses "filter->rtr &
filter->rtr_mask" for setting the filter mask. It should just be using
filter->rtr_mask, otherwise filters for non-RTR frames will match RTR
frames as well.

When transmitting a RTR frame, the hardware automatically switches the
mailbox used for TX to RX in order to receive the reply. This, however,
does not match the Zephyr CAN driver model, where mailboxes are dedicated
to either RX or TX. Attempting to reuse the TX mailbox (which was
automatically switched to an RX mailbox by the hardware) fails on the first
call, after which the mailbox is reset and can be reused for TX. To
overcome this, the driver must abort the RX mailbox operation when the
hardware performs the TX to RX switch.

(cherry picked from commit 3e5aaf837e723b4ef36ad83e86fd2dc56b39f442)

Original-Fixes: #47902
Original-Signed-off-by: Henrik Brix Andersen <hebad@vestas.com>
GitOrigin-RevId: 3e5aaf837e723b4ef36ad83e86fd2dc56b39f442
Change-Id: I5feecb9ce49eed1910d337bf011566828da5743d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3772266
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Aaron Massey <aaronmassey@google.com>